# apps/marketplace/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Marketplace Simulator API
  version: 1.0.0
  description: |
    A provider-aware marketplace API. Order and fulfillment contracts are only
    available through `SHOPEE_LIKE` and `TOKOPEDIA_LIKE`; Generic is not an
    order-provider profile.

    Shared catalogue, warehouse, and webhook requests under `/api/v1` are HMAC-SHA256 signed using `METHOD + PATH + UNIX_TIMESTAMP + RAW_BODY`.
    Every mutating public operation across the shared, Shopee-like, and
    Tokopedia-like contracts require an `Idempotency-Key`. Concurrent use is
    claimed atomically, completed responses are replayed, and reusing a key for
    a different logical request returns `409`.

    Successful and rate-limited public responses expose `X-RateLimit-Limit`,
    `X-RateLimit-Remaining`, and `X-RateLimit-Reset` (Unix seconds). Public requests can return
    `503 MARKETPLACE_MAINTENANCE` when the administrator enables global maintenance mode.

    Webhooks are asynchronous. A 2xx response succeeds; all other responses and network failures
    retry at 30 seconds, 2 minutes, 10 minutes, and 30 minutes. Delivery attempts are immutable.
    Per-shop scenarios can inject slow responses, random 500s, timeouts, rate limits, duplicate,
    delayed, out-of-order, and failed webhook deliveries without affecting other shops.

    `SHOPEE_LIKE` shops additionally expose `/api/shopee/v1`. This is a distinct
...
    `X-Shopee-Signature`, where the signature is HMAC-SHA256 of
...

...
servers:
...
paths:
...
  /control/v1/shops/{id}/products:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string }, description: Shop id. }
    get:
      operationId: listControlProducts
      tags: [Control Plane, Products]
      summary: List products managed by a shop owner or administrator
      security: [{ ControlPlaneBearer: [] }]
      responses:
        '200': { description: Paginated control-plane catalogue. }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
    post:
      operationId: createControlProduct
      tags: [Control Plane, Products]
      summary: Create a product and its initial warehouse inventory
      security: [{ ControlPlaneBearer: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ControlProductInput' } } }
      responses:
        '201': { description: Product created., content: { application/json: { schema: { $ref: '#/components/schemas/ControlProduct' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /control/v1/shops/{id}/products/{productID}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string }, description: Shop id. }
      - { name: productID, in: path, required: true, schema: { type: string } }
    get:
      operationId: getControlProduct
      tags: [Control Plane, Products]
      summary: Get one active control-plane product
      security: [{ ControlPlaneBearer: [] }]
      responses:
        '200': { description: Product detail., content: { application/json: { schema: { $ref: '#/components/schemas/ControlProduct' } } } }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
    patch:
      operationId: updateControlProduct
      tags: [Control Plane, Products]
      summary: Update catalogue metadata; adjust physical stock through warehouse inventory
      security: [{ ControlPlaneBearer: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ControlProductPatch' } } }
      responses:
        '200': { description: Product updated., content: { application/json: { schema: { $ref: '#/components/schemas/ControlProduct' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
    delete:
      operationId: archiveControlProduct
      tags: [Control Plane, Products]
      summary: Archive an unreserved product while preserving order history
      security: [{ ControlPlaneBearer: [] }]
      responses:
        '204': { description: Product archived. }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    ControlPlaneBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-session-token
...
  schemas:
...
    WarehouseInventoryInput:
      type: object
      required: [warehouse_id, on_hand_quantity]
      properties:
        warehouse_id: { type: string }
        on_hand_quantity: { type: integer, minimum: 0 }
    ControlProductInput:
      type: object
      required: [sku, name, price, warehouse_inventory]
      properties:
        sku: { type: string, minLength: 1 }
        name: { type: string, minLength: 1 }
        category: { type: string }
        description: { type: string }
        price: { type: integer, format: int64, minimum: 0, description: Minor currency unit. }
        status: { type: string, enum: [ACTIVE, INACTIVE], default: ACTIVE }
        warehouse_inventory: { type: array, minItems: 1, items: { $ref: '#/components/schemas/WarehouseInventoryInput' } }
    ControlProductPatch:
      type: object
      description: Partial catalogue metadata update. Omitted fields remain unchanged; physical inventory is intentionally excluded.
      properties:
        name: { type: string, minLength: 1 }
        category: { type: string, minLength: 1 }
        description: { type: string }
        price: { type: integer, format: int64, minimum: 0 }
        status: { type: string, enum: [ACTIVE, INACTIVE] }
    ControlProduct:
      type: object
      required: [id, shop_id, sku, name, category, description, price, stock, status]
      properties:
        id: { type: string }
        shop_id: { type: string }
        sku: { type: string }
        name: { type: string }
        category: { type: string }
        description: { type: string }
        price: { type: integer, format: int64 }
        stock: { type: integer, description: Aggregate available quantity across warehouses. }
        status: { type: string, enum: [ACTIVE, INACTIVE] }
...
    Error:
      type: object
      properties: { error: { type: object, properties: { code: { type: string }, message: { type: string } } } }
  responses:
...
    Error: { description: API error, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...